requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
PySocks>=1.7.1
beautifulsoup4>=4.9.3
flask>=3.0.0
flask-socketio>=5.5.0
eventlet>=0.33.0 